backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
groq>=0.13.0
sqlparse==0.4.4
cryptography==42.0.2
python-dotenv==1.0.1
aiofiles==23.2.1
requests==2.31.0
